# docs/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Workflow Execution Service
  version: "1.0.0"
  description: >
    Runtime control plane for the BPMN-driven Workflow Engine. Manages workflow instances,
    task projections for dashboard queries, and user signals (claim, complete, defer, reassign).
    All identity and tenant context is injected by the Envoy API Gateway — callers never set
    these headers directly.
servers:
  - url: https://api.internal/execution
    description: Internal API gateway base (service prefix only; paths include /api/v1/)
security:
  - GatewayTenantContext: []
    UserContext: []
...
components:
  securitySchemes:
    GatewayTenantContext:
      type: apiKey
      in: header
      name: x-tenant-id
      description: Tenant UUID injected by the API Gateway.
    UserContext:
      type: apiKey
      in: header
      name: x-user-id
      description: Authenticated user UUID injected by the API Gateway.
    AdminRole:
      type: apiKey
      in: header
      name: x-tenant-roles
      description: >
        Comma-separated tenant roles injected by the API Gateway. Admin-only endpoints
        require at least one of: tenant_admin, tenant_owner.
    InternalServiceAuth:
      type: apiKey
      in: header
      name: x-internal-token
      description: >
        Service-to-service credential for the internal route group (RequireInternalToken,
        INTERNAL_API_TOKEN) — the same convention Definition Service already ships. These
        routes are NOT exposed through the Envoy public gateway and do NOT carry gateway
        identity headers; they are restricted to in-mesh callers by a Kubernetes
        NetworkPolicy AND authenticated by this shared secret — defense-in-depth, not either
        control alone (§9.2). The tenant context for these routes is taken from the request
        payload, not from x-tenant-id.
  parameters:
    TenantIdHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID (injected by gateway).
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Authenticated user UUID (injected by gateway).
...
    PlanHeader:
      name: x-plan
      in: header
      required: false
      schema:
        type: string
        enum: [starter, pro, enterprise]
      description: Tenant plan tier (gateway-injected).
...
    RequestIdHeader:
      name: x-request-id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Optional client-generated request ID for distributed tracing correlation.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: >
        Optional client-supplied key for safe retry of this request. A replay with the same
        key and an identical request body returns the original cached response. A replay with
        the same key and a different body returns 409 IDEMPOTENCY_KEY_REPLAY.
...
paths:
...
  /api/v1/instances:
    post:
      operationId: startInstance
      tags: [Instances]
      summary: Start a new workflow instance
      description: >
        Validates that the referenced version is PUBLISHED and is_valid=true, then starts
        a Temporal workflow execution and inserts the instance record transactionally.
        Returns 409 if an active instance (RUNNING or PAUSED) with the same business_key
        already exists for this tenant.
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/PlanHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/TraceParentHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [business_key, workflow_version_id]
              properties:
                business_key:
                  type: string
                  minLength: 2
                  maxLength: 64
                  description: >
                    Caller-supplied unique key for this run. Must be unique among
                    RUNNING and PAUSED instances for this tenant.
                  example: "tender-abc-001"
                workflow_version_id:
                  type: string
                  format: uuid
                  description: Must reference a PUBLISHED version with is_valid=true.
                context_json:
                  type: object
                  additionalProperties: true
                  description: >
                    Optional runtime data attached to this instance (e.g. tender_id, document refs).
                    Immutable after creation; passed to all tasks created for this instance.
                override_map:
                  $ref: "#/components/schemas/OverrideMap"
      responses:
        "201":
          description: Instance created and Temporal workflow started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: >
            State conflict. See code field:
            IDEMPOTENCY_KEY_REPLAY — same key reused with a different request body;
            DUPLICATE_BUSINESS_KEY — an active instance for this business_key already exists;
            VERSION_NOT_PUBLISHED — the referenced version is DRAFT or ARCHIVED;
            VERSION_INVALID — version.is_valid=false (assignee eligibility was revoked);
            TENANT_NOT_ACTIVE — the tenant's current status is not active or trial.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                DuplicateBusinessKey:
                  value:
                    type: "https://errors.bcbp.io/execution/duplicate-business-key"
                    title: "Duplicate Business Key"
                    status: 409
                    detail: "an active instance for business_key 'tender-abc-001' already exists"
                    instance: "/api/v1/instances"
                    code: "DUPLICATE_BUSINESS_KEY"
                TenantNotActive:
                  value:
                    type: "https://errors.bcbp.io/execution/tenant-not-active"
                    title: "Tenant Not Active"
                    status: 409
                    detail: "tenant status is 'suspended'; instantiation requires active or trial"
                    instance: "/api/v1/instances"
                    code: "TENANT_NOT_ACTIVE"
                VersionNotPublished:
                  value:
                    type: "https://errors.bcbp.io/execution/version-not-published"
                    title: "Version Not Published"
                    status: 409
                    detail: "workflow version is not PUBLISHED"
                    instance: "/api/v1/instances"
                    code: "VERSION_NOT_PUBLISHED"
                VersionInvalid:
                  value:
                    type: "https://errors.bcbp.io/execution/version-invalid"
                    title: "Version Invalid"
                    status: 409
                    detail: "workflow version is_valid=false; default assignee eligibility was revoked"
                    instance: "/api/v1/instances"
                    code: "VERSION_INVALID"
                IdempotencyReplay:
                  value:
                    type: "https://errors.bcbp.io/execution/idempotency-key-replay"
                    title: "Idempotency Key Replay"
                    status: 409
                    detail: "Idempotency-Key reused with a different request body"
                    instance: "/api/v1/instances"
                    code: "IDEMPOTENCY_KEY_REPLAY"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          description: >
            See code field: OVERRIDE_MAP_INVALID — override_map contains unknown node keys or
            invalid UUIDs; ASSIGNEE_INELIGIBLE — a default assignee no longer satisfies their
            node's (department, level) eligibility requirement (§5.5).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                OverrideMapInvalid:
                  value:
                    type: "https://errors.bcbp.io/execution/override-map-invalid"
                    title: "Override Map Invalid"
                    status: 422
                    detail: "node key 'unknown_stage' does not exist in the compiled plan"
                    instance: "/api/v1/instances"
                    code: "OVERRIDE_MAP_INVALID"
                AssigneeIneligible:
                  value:
                    type: "https://errors.bcbp.io/execution/assignee-ineligible"
                    title: "Assignee Ineligible"
                    status: 422
                    detail: "one or more default assignees no longer satisfy their node's eligibility requirement"
                    instance: "/api/v1/instances"
                    code: "ASSIGNEE_INELIGIBLE"
                    invalid_params:
                      - name: "review_finance"
                        reason: "user no longer holds required_level=SENIOR in department_id=9b1c...finance"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    get:
      operationId: listInstances
      tags: [Instances]
      summary: List workflow instances
      description: >
        Returns a paginated list of instances for the caller's tenant, additionally scoped to
        the same assignee/department/admin visibility rule as GET /instances/:id (§9.2) —
        non-visible instances are simply absent from the page, not an error.
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/InstanceStatus"
          description: Filter by instance status.
        - name: workflow_version_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter to a specific compiled version.
        - name: started_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Include only instances started after this timestamp (inclusive).
        - name: started_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Include only instances started before this timestamp (exclusive).
      responses:
        "200":
          description: Keyset page of workflow instances (§5.9).
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/InstanceSummary"
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for the next page; omitted/null on the final page.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/instances/{id}:
    get:
      operationId: getInstance
      tags: [Instances]
      summary: Get workflow instance detail
      description: >
        Returns instance detail including all tasks created for this instance. Visible only
        if the caller is (or was) an assignee on any task under this instance, the caller's
        x-departments includes any currently-open task's department_id, or the caller holds
        tenant_admin/tenant_owner (§9.2). Returns 403 NOT_AUTHORIZED_FOR_RESOURCE otherwise —
        not 404, since the instance is already confirmed to exist in the caller's own tenant.
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantRolesHeader"
        - $ref: "#/components/parameters/DepartmentsHeader"
        - $ref: "#/components/parameters/InstanceIdPath"
      responses:
        "200":
          description: Workflow instance detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: >
            NOT_AUTHORIZED_FOR_RESOURCE — caller is not an assignee, not in the current
            department, and not an admin.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/not-authorized-for-resource"
                title: "Not Authorized For Resource"
                status: 403
                detail: "caller is not an assignee, department member, or admin for this instance"
                instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6"
                code: "NOT_AUTHORIZED_FOR_RESOURCE"
        "404":
          $ref: "#/components/responses/InstanceNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/instances/{id}/pause:
    post:
      operationId: pauseInstance
      tags: [Instances]
      summary: Pause a running workflow instance (admin)
      description: >
        Sends a pause signal to the Temporal workflow. The instance transitions from
        RUNNING to PAUSED. Returns 409 if the instance is not RUNNING.
      security:
        - GatewayTenantContext: []
          UserContext: []
          AdminRole: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantRolesHeader"
        - $ref: "#/components/parameters/InstanceIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_version]
              properties:
                record_version:
                  type: integer
                  description: >
                    The caller's current view of the instance's record_version, checked
                    synchronously before the signal is forwarded (§5.10).
                  example: 3
                reason:
                  type: string
                  maxLength: 500
                  nullable: true
                  description: Optional human-readable reason for the pause.
      responses:
        "202":
          description: Pause signal accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/InstanceNotFound"
        "409":
          description: >
            State conflict. See code field: INVALID_INSTANCE_STATE — instance is not RUNNING;
            RECORD_VERSION_CONFLICT — record_version is stale;
            IDEMPOTENCY_KEY_REPLAY — same key reused with a different request body.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                InstanceStateInvalid:
                  value:
                    type: "https://errors.bcbp.io/execution/invalid-instance-state"
                    title: "Invalid Instance State"
                    status: 409
                    detail: "instance is not RUNNING"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/pause"
                    code: "INVALID_INSTANCE_STATE"
                RecordVersionConflict:
                  value:
                    type: "https://errors.bcbp.io/execution/record-version-conflict"
                    title: "Record Version Conflict"
                    status: 409
                    detail: "the instance has been modified since record_version=3 was read"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/pause"
                    code: "RECORD_VERSION_CONFLICT"
                IdempotencyReplay:
                  value:
                    type: "https://errors.bcbp.io/execution/idempotency-key-replay"
                    title: "Idempotency Key Replay"
                    status: 409
                    detail: "Idempotency-Key reused with a different request body"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/pause"
                    code: "IDEMPOTENCY_KEY_REPLAY"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/instances/{id}/resume:
    post:
      operationId: resumeInstance
      tags: [Instances]
      summary: Resume a paused workflow instance (admin)
      description: >
        Sends a resume signal to the Temporal workflow. The instance transitions from
        PAUSED to RUNNING. Returns 409 if the instance is not PAUSED.
      security:
        - GatewayTenantContext: []
          UserContext: []
          AdminRole: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantRolesHeader"
        - $ref: "#/components/parameters/InstanceIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_version]
              properties:
                record_version:
                  type: integer
                  description: >
                    The caller's current view of the instance's record_version, checked
                    synchronously before the signal is forwarded (§5.10).
                  example: 4
      responses:
        "202":
          description: Resume signal accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/InstanceNotFound"
        "409":
          description: >
            State conflict. See code field: INVALID_INSTANCE_STATE — instance is not PAUSED;
            RECORD_VERSION_CONFLICT — record_version is stale;
            IDEMPOTENCY_KEY_REPLAY — same key reused with a different request body.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                InstanceStateInvalid:
                  value:
                    type: "https://errors.bcbp.io/execution/invalid-instance-state"
                    title: "Invalid Instance State"
                    status: 409
                    detail: "instance is not PAUSED"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/resume"
                    code: "INVALID_INSTANCE_STATE"
                RecordVersionConflict:
                  value:
                    type: "https://errors.bcbp.io/execution/record-version-conflict"
                    title: "Record Version Conflict"
                    status: 409
                    detail: "the instance has been modified since record_version=4 was read"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/resume"
                    code: "RECORD_VERSION_CONFLICT"
                IdempotencyReplay:
                  value:
                    type: "https://errors.bcbp.io/execution/idempotency-key-replay"
                    title: "Idempotency Key Replay"
                    status: 409
                    detail: "Idempotency-Key reused with a different request body"
                    instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/resume"
                    code: "IDEMPOTENCY_KEY_REPLAY"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...